requirements.txt
ipython
falcon
gunicorn
msgpack-python
requests
httpie
mongoengine
bcrypt
PyJWT